refactor layout to avoid useEffect in server component
